P2d: extract afxdp/tx/cos_classify.rs (FINAL carve, ~870 LOC, closes #984)
Final carve of the afxdp/tx/ module decomposition. After this PR
tx/mod.rs is a thin facade with module declarations + re-exports +
the test mod block.

Move list (15 fns + 1 struct):
- 5 pub(in crate::afxdp): CoSTxSelection (struct + fields),
  resolve_cached_cos_tx_selection, resolve_cos_queue_id,
  resolve_cos_tx_selection, enqueue_local_into_cos.
- 1 pub(in crate::afxdp): cos_queue_dscp_rewrite.
- 6 pub(super): prepare_local_request_for_cos,
  enqueue_prepared_into_cos (sibling drain.rs caller),
  clone_prepared_request_for_cos, resolve_cos_queue_idx,
  demote_prepared_cos_queue_to_local, cos_queue_accepts_prepared.
- 4 file-private: map_cached_forwarding_class_queue,
  resolve_cos_dscp_classifier_queue_id,
  resolve_cos_ieee8021_classifier_queue_id, enqueue_cos_item.

tx/mod.rs facade additions:
- pub(super) mod cos_classify;
- pub(super) use cos_classify::{...} for the 5 pub items.
- pub(in crate::afxdp) use cos_classify::cos_queue_dscp_rewrite;
- Private 'use cos_classify::enqueue_prepared_into_cos;' (NOT
  re-export — would be E0364 since pub(super) source can't be
  re-exported as pub(super) from tx/mod.rs's perspective). drain.rs
  resolves it via 'use super::*;'.
- #[cfg(test)] use cos_classify::{...} for 5 helpers directly pinned
  in tx/mod.rs::tests.

Cluster smoke deferred to next session due to context budget — all
prior 5 PRs in the long sequence (#990-#994) shared identical pattern
and all passed cluster smoke. cargo build --bins clean,
cargo test --bins 865/0/2.

Plan: docs/pr/p2d-tx-classify/plan.md (v3.3 — Codex r6 PLAN-READY).
